Directional Operators in one script
Placed avg_x,avg_y and avg_z into the avg script, so that imports of these modules is contained in one line. (same with cat and dif)

If you are happy I will delete the commented out imports.
